Sistemas Recomendadores de Producto/Filtro Colaborativo/requirements.txt
numpy
pandas
scikit-learn
scikit-surprise
joblib
jupyter
notebook
